requirements.txt
asn1crypto==1.5.1
awscli==1.24.1
beautifulsoup4==4.11.1
boto3==1.23.1
botocore==1.26.1
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dataclasses==0.6
docutils==0.15.2
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
lxml==4.8.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
redshift-connector==2.0.907
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
scramp==1.4.1
six==1.16.0
soupsieve==2.3.2.post1
troposphere==4.0.2
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
